Adding source code.
This is the finalised build, which was submitted for my final year dissertation project. 

It is far from complete and still needs work doing to it, which I may get round to in time. However, some people may gain something from the project.  It is functional and was brought online for a period of time to gather results for the report/project.

- snoom